moving histogram pattern to histogram_impl
